Restructuring of namelists and SWIO output linking.
Namelists and resource files are now stored in the PARMDIR
(scripts/compsets/parm), and are envsubst'd into the RUNDIR as needed.

SWIO output is linked out as needed. It requires the user to define
SWIO=.true., and the SWIO_MODELS="IO AIO" etc. In addition, the user
must define $IO_CADENCE and $IO_PREFIX ($AIO_CADENCE, $AIO_PREFIX,
etc.) in order to properly link the output from the RUNDIR to the
ROTDIR.

I ran through some options for automatically grabbing this information
from the appropriate template nems.configure and the SWIO resource
files, but every solution wound up being too convoluted for quick
implementation. I think, ultimately, a Python script that parses the
appropriate files in the PARMDIR will be the solution, but it's not
there yet.