Realtime Input Parameters overhaul
In preparation for CONOPS2, an overhaul of the realtime input
parameters parser was performed. Missing values are interpolated
linearly, while forecasted values relax with a time constant to
values calculated by F10.7 and Kp forecasts from SWFO.

Output can now be sub-hour, although the averaging for solar wind values
will need to be adjusted to allow for accurate calculation of those
values in averaging windows longer than the output interval.

The compsets workflow has not been updated to take advantage of these
features as of yet. This will need to be merged with the conops2 branch,
and the WAM and IPE components updated to re-read the wam_input_f107_kp.txt
file when necessary. Further, we'll need a wrapper around the input parameter
generation that can interface with the DATAPOLL component appropriately.