Support for driver end dates.
Currently the realtime driver parser is as greedy as possible;
this is an issue when running historical operational runs if the
goal is to replicate all data that was available to that run when
it was executing. This introduces attributes to the NetCDF driver file
that store those values, and adds functionality that allows those
values to be passed, such that you can recreate a given driver file
with a driver database that contains more data.