Adding some new options to the driver parsers.
I exercised severe restraint in not totally rewriting the historical
parser; that will have to wait for another time.

Main user-facing changes: exposing options in user.config called
REALTIME_DERIVE and HISTORICAL_NEW_F107 allowing for the user to run
with the realtime driver database but use Kp-derived drivers and opt
for the new, observed F10.7 database, which I hope to transition to
exclusively going forward. Examples in the march2013_regression.config.