F10.7 handling changes
Both the historical and realtime parser now have their F10.7 observed
times set to 20UT which is consistent with the new F10.7 database
we intended to use as well as the values provided by SWFO for
operations.

Additionally, functionality has been added to the realtime parser to
not allow unrealistic F10.7 and F10.7a jumps, with the intent of
adding resiliency to the operational system.